tried to add records to the DB

+++ internals/sql/schema/002_prices.sql
@@ -2,7 +2,7 @@
 CREATE TABLE Prices (
     PriceDate TIMESTAMP,
     Item_id INTEGER NOT NULL REFERENCES Items(id),
-    Price FLOAT,
+    Price NUMERIC(10, 2),
     PRIMARY KEY(Item_id, PriceDate)
 );
 -- +goose Down
